requirements.txt
torch>=2.1
torchvision>=0.16
numpy>=1.24
Pillow>=9.0
Flask>=2.3
pix2tex
sympy
langchain_google_genai
dotenv
<<<<<<< HEAD
pip install antlr4-python3-runtime==4.11 sympy
=======

>>>>>>> 4a199d77ee220a94d59fd8e958a07d76fdaf495d


